MNT-23915: errors thrown in called activities get caught by parent process (#4490)
* MNT-22737: Handle Boundary Event Properly (#3947)

* MNT-23915: errors thrown in called activities get caught by parent process (#4474)
* Only checks if it's a called activity if there's no catch been executed
(cherry picked from commit 3c2e714f96e8949599835c703a4fd5d92b8bc3b7)

+++ activiti-core/activiti-engine/src/main/java/org/activiti/engine/impl/bpmn/behavior/InclusiveGatewayActivityBehavior.java
@@ -69,22 +69,21 @@ public class InclusiveGatewayActivityBehavior extends GatewayActivityBehavior im
 
     Collection<ExecutionEntity> allExecutions = executionEntityManager.findChildExecutionsByProcessInstanceId(execution.getProcessInstanceId());
     Iterator<ExecutionEntity> executionIterator = allExecutions.iterator();
-    boolean oneExecutionCanReachGateway = false;
-    while (!oneExecutionCanReachGateway && executionIterator.hasNext()) {
+    boolean oneExecutionCanReachGatewayInstance = false;
+    while (!oneExecutionCanReachGatewayInstance && executionIterator.hasNext()) {
       ExecutionEntity executionEntity = executionIterator.next();
       if (!executionEntity.getActivityId().equals(execution.getCurrentActivityId())) {
-        boolean canReachGateway = ExecutionGraphUtil.isReachable(execution.getProcessDefinitionId(), executionEntity.getActivityId(), execution.getCurrentActivityId());
-        if (canReachGateway) {
-          oneExecutionCanReachGateway = true;
-        }
-      } else if (executionEntity.getActivityId().equals(execution.getCurrentActivityId()) && executionEntity.isActive()) {
+          if (isSameExecutionPath(execution, executionEntity) && ExecutionGraphUtil.isReachable(execution.getProcessDefinitionId(), executionEntity.getActivityId(), execution.getCurrentActivityId())) {
+              oneExecutionCanReachGatewayInstance = true;
+          }
+      } else if (executionEntity.getId().equals(execution.getId()) && executionEntity.isActive()) {
         // Special case: the execution has reached the inc gw, but the operation hasn't been executed yet for that execution
-        oneExecutionCanReachGateway = true;
+        oneExecutionCanReachGatewayInstance = true;
       }
     }
 
     // If no execution can reach the gateway, the gateway activates and executes fork behavior
-    if (!oneExecutionCanReachGateway) {
+    if (!oneExecutionCanReachGatewayInstance) {
 
       logger.debug("Inclusive gateway cannot be reached by any execution and is activated");
 
@@ -92,14 +91,18 @@ public class InclusiveGatewayActivityBehavior extends GatewayActivityBehavior im
       Collection<ExecutionEntity> executionsInGateway = executionEntityManager
           .findInactiveExecutionsByActivityIdAndProcessInstanceId(execution.getCurrentActivityId(), execution.getProcessInstanceId());
       for (ExecutionEntity executionEntityInGateway : executionsInGateway) {
-        if (!executionEntityInGateway.getId().equals(execution.getId())) {
-          commandContext.getHistoryManager().recordActivityEnd(executionEntityInGateway, null);
-          executionEntityManager.deleteExecutionAndRelatedData(executionEntityInGateway, null);
-        }
+          if (!executionEntityInGateway.getId().equals(execution.getId()) && executionEntityInGateway.getParentId().equals(execution.getParentId())) {
+              commandContext.getHistoryManager().recordActivityEnd(executionEntityInGateway, null);
+              executionEntityManager.deleteExecutionAndRelatedData(executionEntityInGateway, null);
+          }
       }
 
       // Leave
       commandContext.getAgenda().planTakeOutgoingSequenceFlowsOperation(execution, true);
     }
+  }
+
+  private boolean isSameExecutionPath(ExecutionEntity gatewayExecution, ExecutionEntity activeExecution) {
+        return activeExecution.getParentId().equals(gatewayExecution.getParentId());
   }
 }
